Extract support preference components module
