+++ rxlifecycle-components-preference/gradle.properties
@@ -1,0 +1,4 @@
+POM_NAME=RxLifecycle-Components-Preference
+POM_DESCRIPTION=RxLifecycle-Components-Preference
+POM_ARTIFACT_ID=rxlifecycle-components-preference
+POM_PACKAGING=aar
