+++ settings.gradle
@@ -2,6 +2,7 @@ include ':rxlifecycle', ':rxlifecycle-android-lifecycle-kotlin'
 include ':rxlifecycle-android'
 include ':rxlifecycle-android-lifecycle'
 include ':rxlifecycle-components'
+include ':rxlifecycle-components-preference'
 include ':rxlifecycle-navi'
 include ':rxlifecycle-kotlin'
 include ':rxlifecycle-sample'
